fix(playback): extract maybe_broadcast_position_update to keep mod.rs under 1000 lines
Phase 2+3 additions pushed playback/mod.rs over the 1000-line file-size cap
(was 1064). Three targeted extractions:

1. position_update.rs — maybe_broadcast_position_update() moved to its own
   file with an impl super::PlaybackEngine { ... } block. Saves 83 lines.
2. db/models.rs — new get_video_suppress_resolume_en(pool, video_id)
   helper replaces a 9-line inline sqlx::query_scalar + filter chain in
   the playback DB-fetch site.
3. presenter/mod.rs — maybe_push_line() helper replaces the 34-line
   inline Presenter push block in maybe_broadcast_position_update (now
   in position_update.rs — still counts against its file-size).

Also: shorter doc comments on three new PlaybackEngine fields, and
handlers.rs gets a blank line inserted in the doc comment to silence
clippy::doc_lazy_continuation ("doc list item without indentation").

Plus the earlier lint+test fix: lib_tests.rs's `mod tests { use super::*; }`
needed `use crate::*;` because the tests moved into a child module.
